refactor: move markdown-table from core/lib/actions to core/lib/general
