+++ core/lib/general/markdown-table.ts
